refactor: unificação controllers
