+++ public/index.php
@@ -3,10 +3,7 @@
 declare(strict_types=1);
 
 use ControleAlunos\Router;
-use ControleAlunos\Controllers\RegistroEscolasController;
-use ControleAlunos\Controllers\ListagemEscolasController;
-use ControleAlunos\Controllers\AtualizacaoEscolasController;
-use ControleAlunos\Controllers\ObterDadosEscolaController;
+use ControleAlunos\Controllers\EscolasController;
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
@@ -14,16 +11,19 @@ $dotenv = Dotenv\Dotenv::createUnsafeImmutable(__DIR__ . "/../");
 $dotenv->safeLoad();
 
 $routes = [
-    '' => ListagemEscolasController::class,
-    'registrar-escola' => RegistroEscolasController::class,
-    'listagem-escolas' => ListagemEscolasController::class,
-    'atualizar-escola' => AtualizacaoEscolasController::class,
-    'obter-dados-escola' => ObterDadosEscolaController::class
+    'GET|/' => [EscolasController::class, 'index'],
+    'GET|/registrar-escola' => [EscolasController::class, 'index'],
+    'POST|/registrar-escola' => [EscolasController::class, 'store'],
+    'GET|/listagem-escolas' => [EscolasController::class, 'index'],
+    'GET|/atualizar-escola' => [EscolasController::class, 'index'],
+    'POST|/atualizar-escola' => [EscolasController::class, 'update'],
+    'POST|/obter-dados-escola' => [EscolasController::class, 'find']
 ];
 
-$request = $_SERVER['REQUEST_URI'];
+$pathInfo = $_SERVER['PATH_INFO'] ?? '/';
+$httpMethod = $_SERVER['REQUEST_METHOD'];
 
-$parsed_uri = parse_url($request);
+$controllerData = $routes["$httpMethod|$pathInfo"];
 
 $database = new PDO("mysql:host=" . getenv("DATABASE_HOST") . 
                ";dbname=" . getenv("DATABASE_DB"), 
@@ -31,4 +31,4 @@ $database = new PDO("mysql:host=" . getenv("DATABASE_HOST") .
                getenv("DATABASE_PASSWORD"));
 
 $router = new Router($routes, ['database' => $database]);
-$router->run($parsed_uri['path']);
+$router->run($controllerData);
